split dashboard code
