+++ dashboard/src/index.css
@@ -183,7 +183,7 @@ a[alt]:hover:after {
 }
 .channel {
 	padding: 0 8px;
-	margin: 0 8px 2px;
+	margin: 0 8px 2px 12px;
 	height: 32px;
 	border-radius: 4px;
 	display: flex;
@@ -193,17 +193,36 @@ a[alt]:hover:after {
 .channel:hover {
 	background: rgba(79,84,92,0.16);
 }
-.channel svg {
+.channel img {
 	margin-right: 6px;
-	width: 20;
-	height: 20;
+	width: 20px;
+	height: 20px;
 }
 .channel div {
 	font-size: 16px;
 	line-height: 20px;
 	text-shadow: 0 0 1px;
 	white-space: nowrap;
+	overflow: hidden;
 }
 .channel:hover div {
 	color: #dcddde;
+}
+.channel.selected div {
+	color: #ffffff;
+}
+.channel-header {
+	margin-left: 8px;
+	height: 44px;
+}
+.channel-header svg {
+	margin-right: 6px;
+	width: 24px;
+	height: 24px;
+}
+.channel-header div {
+	font-size: 20px;
+	line-height: 24px;
+	font-weight: bold;
+	text-shadow: none;
 }
